refactor(ui): some ui structure refactor
some ui structure refactor
