+++ packages/ui/src/App.tsx
@@ -1,9 +1,7 @@
-import { conditionalString } from '@silverhand/essentials';
 import React, { useEffect } from 'react';
 import { Route, Routes, BrowserRouter, Navigate } from 'react-router-dom';
 
-import * as styles from './App.module.scss';
-import AppContent from './components/AppContent';
+import AppContent from './containers/AppContent';
 import LoadingLayerProvider from './containers/LoadingLayerProvider';
 import usePageContext from './hooks/use-page-context';
 import usePreview from './hooks/use-preview';
@@ -29,8 +27,6 @@ const App = () => {
 
   useEffect(() => {
     if (isPreview) {
-      document.body.classList.add(conditionalString(styles.preview));
-
       return;
     }
 
@@ -61,17 +57,21 @@ const App = () => {
             />
 
             <Route element={<LoadingLayerProvider />}>
-              {/* always keep route path with param as the last one */}
               <Route path="/sign-in" element={<SignIn />} />
               <Route path="/sign-in/:method" element={<SecondarySignIn />} />
               <Route path="/register" element={<Register />} />
               <Route path="/register/:method" element={<Register />} />
 
               {/* social sign-in pages */}
-              <Route path="/sign-in/callback/:connector" element={<SocialSignInCallback />} />
+              <Route
+                path="/social/sign-in-callback/:connector"
+                element={<SocialSignInCallback />}
+              />
               <Route path="/callback/:connector" element={<Callback />} />
               <Route path="/social/register/:connector" element={<SocialRegister />} />
               <Route path="/social/landing/:connector" element={<SocialLanding />} />
+
+              {/* always keep route path with param as the last one */}
               <Route path="/:type/:method/passcode-validation" element={<Passcode />} />
             </Route>
